requirements.txt
Django==5.1.4
dnspython==2.7.0
feedparser==6.0.11
ioc_finder==7.3.0
lxml==5.3.0
Requests==2.32.3
flower
celery
django-celery-results
django-celery-beat
dj_database_url 
psycopg2-binary
django-redis
gunicorn
whitenoise
xmltodict
whois
djangorestframework
markdown
django-filter
tiktoken
google.generativeai
tldextract
django-cors-headers
stix2
PyYAML
python-dotenv
google-adk